requirements.txt
alembic==1.15.1
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dotenv==0.9.9
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.0.1
SQLAlchemy==2.0.38
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
